fit lavam fuzzing to aflplusplus & path_fuzzer_full_path

+++ targets/uniq/build.sh
@@ -16,26 +16,35 @@ if [[ "$AFL_LLVM_CMPLOG" != "1" ]]; then
     export CFGFILE="$OUT/cfg.txt"
     make -e cyh$program
 
-    cat $OUT/cfg.txt | grep "BasicBlock: " | wc -l > $OUT/bbnum.txt
-    # 如果是其它 TARGET，这里需要一个 filter CFG 和 callmap 的操作，这里直接拷贝改名就好
-    cp $OUT/cfg.txt $OUT/cfg_filtered.txt
-    cp $OUT/callmap.txt $OUT/callmap_filtered.txt
-    # 生成 function_list.txt
-    cat $OUT/cfg_filtered.txt | grep "Function: " | nl -v 0 | awk '{print $1, $3, $4, $5, $6, $7, $8, $9}' > $OUT/function_list.txt
+    if [[ -e "$OUT/cfg.txt" ]] && [[ -e "$OUT/callmap.txt" ]]; then
+        cat $OUT/cfg.txt | grep "BasicBlock: " | wc -l > $OUT/bbnum.txt
+        # 如果是其它 TARGET，这里需要一个 filter CFG 和 callmap 的操作，这里直接拷贝改名就好
+        cp $OUT/cfg.txt $OUT/cfg_filtered.txt
+        cp $OUT/callmap.txt $OUT/callmap_filtered.txt
+        # 生成 function_list.txt
+        cat $OUT/cfg_filtered.txt | grep "Function: " | nl -v 0 | awk '{print $1, $3, $4, $5, $6, $7, $8, $9}' > $OUT/function_list.txt
 
-    cp "src/$program" "$OUT/afl/$program"
+        cp "src/$program" "$OUT/afl/$program"
 
-    g++ -g $FUZZER/convert.cpp -o $OUT/convert
-    pushd $OUT > /dev/null
-    ./convert
-    popd > /dev/null
+        g++ -g $FUZZER/convert.cpp -o $OUT/convert
+        pushd $OUT > /dev/null
+        ./convert
+        popd > /dev/null
 
-    mv $OUT/top.bin $OUT/${program}_cfg.bin
+        mv $OUT/top.bin $OUT/${program}_cfg.bin
 
-    # 改名，防止冲突
-    mv $OUT/cfg_filtered.txt $OUT/cfg_${program}.txt
-    mv $OUT/callmap_filtered.txt $OUT/callmap_${program}.txt
-    mv $OUT/function_list.txt $OUT/${program}_function_list.txt
+        # 改名，防止冲突
+        mv $OUT/cfg_filtered.txt $OUT/cfg_${program}.txt
+        mv $OUT/callmap_filtered.txt $OUT/callmap_${program}.txt
+        mv $OUT/function_list.txt $OUT/${program}_function_list.txt
+    else
+
+        if [[ "$FUZZER" == *"aflplusplus"* ]]; then
+            echo "Assertion failed: Fuzzer is not aflplusplus, but cfg.txt does not exit"
+            exit 1
+        fi
+
+    fi
 
     popd
